Refaktoroi: Uudelleenorganisoi koodia
